Move asset module into separate file
